requirements.txt
python-docx>=0.8.11
markdown>=3.4.4
PyYAML>=6.0
click>=8.1.0
Pillow>=9.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0